requirements.txt
# requirements.txt
numpy==1.21.0
torch==1.9.0
matplotlib==3.4.2
gym==0.18.0
pyyaml==5.4.1
tqdm==4.61.0
pandas==1.3.0